refine example
